Library

Newsvendor Model

NewsvendorModel.NVModel — Type

NVModel(c, p, d)

Define a Newsvendor model with unit cost c, unit selling price p, and demand distribution d. Note that d is assumed to be a univariate distribution from the package Distributions.jl.

Examples

julia> using Distributions
julia> nvm = NVModel(1, 5, Normal(90, 30))

This defines a model with unit cost 1, unit price 5, and a demand distributed that follows normal distribution with mean 90 and standard deviation 30.

Further keyword arguments:

NVModel(c, p, d [; salvage = 0, backlog = 0, fixcost = 0])

salvage = unit salvage value; defaults to 0

backlog = unit backlog penalty (e.g., penalty if not delivered or missed future profit for each unserved customer); defaults to 0

fixcost = fixed cost; defaults to 0

Examples

julia> using Distributions
julia> nvm2 = NVModel(1, 5, Normal(90, 30), salvage = 0.5, backlog = 2, fixcost = 100)

defines a newsvendor problem where c = 1, p = 5, d = Normal(90, 30), the salvage value = 0.5, unit backlog penalty = 2, and fixed cost of the operation = 100.

External links

Newsvendor model on Wikipedia

Optimizing expected profit

NewsvendorModel.optimize — Function

optimize(nvm::NVModel [; rounduprule = true])

Compute the stocking quantity q_opt(nvm) that maximizes the expected profit as well as further the metrics when using it. The optimal quantity is rounded up to the next integer by default. To get the exact real number, set the keyword arguement rounduprule = false.

Abstract NewsvendorProblem

NewsvendorModel.NewsvendorProblem — Type

An abstract newsvendor problem is essentially described by having a

demand distribution,

cost of overage,

cost of underage, and

profitshift term.

Its standard concrete type might come as a NewsvendorModel, defined by unit cost, uni selling price etc.

Required functions that determine a newsvendor problem

NewsvendorModel.overagecost — Method

overage(nvpr::NewsvendorProblem)

Get the cost of overage of a newsvendor problem.

NewsvendorModel.underagecost — Method

underage(nvpr::NewsvendorProblem)

Get the cost of underage of a newsvendor problem.

NewsvendorModel.distr — Method

distr(nvpr::NewsvendorProblem)

Get the demand distribution of a newsvendor problem.

NewsvendorModel.profitshift — Method

profitshift(nvpr::NewsvendorProblem)

Define how profit is shifted because of fixed cost, penalty, etc.; defautls to 0.

Quantities of interest

NewsvendorModel.q_opt — Function

q_opt(nvpr::NewsvendorProblem; rounduprule = true)

Compute the quanitity that maximizes expected profit for a newsvendor problem (i.e., where critical fractile equals in-stock probability). Relies on the equation

\[F(q_{opt}) = \textrm{critical fractile} \]

where F is the c.d.f. of the demand distribution. Returns the next integer unless rounduprule=false; then, it returns exact real.

q_opt(res::NVResult)

Get optimal quantity from a stored result.

NewsvendorModel.q_scarf — Function

q_scarf(nvpr::NewsvendorProblem)

Compute the quanitity that maximizes the minimal expected profit among all distributions with the same mean and variance. Worst-case solution (Scarf, 1958)

\[q_{scarf} = \mu + \sigma/2 * (\sqrt{r} - \sqrt{1/r})\]

where

\[r = \frac{\textrm{underage cost}}{ \textrm{overage cost}}\]

Metrics

NewsvendorModel.criticalfractile — Function

criticalfractile(nvpr::NewsvendorProblem)

Compute the critical fractile for a newsvendor problem:

\[\textrm{critical fractile} = \frac{\textrm{underage cost}}{\textrm{underage cost} + \textrm{overage cost}}\]

criticalfractile(res::NVResult)

Get critical fractile from a stored result.

NewsvendorModel.profit — Function

profit(nvpr::NewsvendorProblem, q)

Compute expected left over quantity when stocking quantity q. It is given by

\[E[\textrm{profit}] = \textrm{underagecost} \times \mu - E[\textrm{missmatch}] + \textrm{profitshift}\]

profit(res::NVResult)

Get expected profit from a stored result.

NewsvendorModel.missmatch — Function

missmatch(nvpr::NewsvendorProblem, q)

Compute expected missmatch when stocking quantity q. It is given by

\[E[\textrm{missmatch cost}] = \textrm{underage cost} \times E[\textrm{lost sales}] + \textrm{overage cost} \times E[\textrm{left over}]\]

NewsvendorModel.lostsales — Function

lostsales(nvpr::NewsvendorProblem, q)

Compute expected lost sales when stocking quantity q.

lostsales(res::NVResult)

Get expected lost sales model from a stored result.

NewsvendorModel.sales — Function

sales(nvpr::NewsvendorProblem, q)

Compute expected sales when stocking quantity q.

sales(res::NVResult)

Get expected sales from a stored result.

NewsvendorModel.leftover — Function

leftover(nvpr::NewsvendorProblem, q)

Compute expected left over quantity when stocking quantity q.

leftover(res::NVResult)

Get expected left over from a stored result.